Moved particle, block, and chunk into their own files (INCOMPLETE and DOESN'T RUN YET)
